Add, delete, update

+++ src/Controller/CourseController.php
@@ -33,20 +33,39 @@ class CourseController extends AbstractController
      * @Route("/new", name="course_new", methods={"GET","POST"})
      * @IsGranted("ROLE_SUPER_ADMIN")
      */
-    public function new(Request $request): Response
+    public function new(Request $request, BillingClient $billingClient): Response
     {
-        $course = new Course();
-        $form = $this->createForm(CourseType::class, $course);
+        $form = $this->createForm(CourseType::class);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($course);
-            $entityManager->flush();
-            return $this->redirectToRoute('course_index');
+            $formData = $form->getData();
+            try {
+                $addResponse = $billingClient->addCourse($formData, $this->getUser()->getApiToken());
+            } catch (HttpException $e) {
+                return $this->render('course/new.html.twig', array(
+                    'form' => $form->createView(),
+                    'error' => "Сервис временно недоступен. Попробуйте добавить урок позднее"
+                ));
+            }
+            if (array_key_exists('code', $addResponse)) {
+                return $this->render('course/new.html.twig', array(
+                'form' => $form->createView(),
+                'error' => $addResponse['message']
+                ));
+            } else {
+                $course = new Course();
+                $course->setName($formData['name']);
+                $course->setDescription($formData['description']);
+
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->persist($course);
+                $entityManager->flush();
+                return $this->redirectToRoute('course_index');
+            }
         }
         return $this->render('course/new.html.twig', [
-            'course' => $course,
             'form' => $form->createView(),
+            'error' => null
         ]);
     }
 
@@ -59,11 +78,13 @@ class CourseController extends AbstractController
         if ($auth_checker->isGranted('ROLE_USER')) {
             return $this->render('course/show.html.twig', [
                 'course' => $courseRepository->findOneCombined($slug, $billingClient, $this->getUser()),
-                'user_balance' => $billingClient->getCurentUserBalance($this->getUser()->getApiToken())
+                'user_balance' => $billingClient->getCurentUserBalance($this->getUser()->getApiToken()),
+                'error' => null
             ]);
         } else {
             return $this->render('course/show.html.twig', [
-                'course' => $courseRepository->findOneCombined($slug, $billingClient, null)
+                'course' => $courseRepository->findOneCombined($slug, $billingClient, null),
+                'error' => null
             ]);
         }
     }
@@ -72,17 +93,40 @@ class CourseController extends AbstractController
      * @Route("/{slug}/edit", name="course_edit", methods={"GET","POST"})
      * @IsGranted("ROLE_SUPER_ADMIN")
      */
-    public function edit(Request $request, Course $course): Response
+    public function edit($slug, Request $request, Course $course, BillingClient $billingClient): Response
     {
-        $form = $this->createForm(CourseType::class, $course);
+        $combinedCourse = $this->getDoctrine()->getRepository(Course::class)->findOneCombined($slug, $billingClient, $this->getUser());
+        $form = $this->createForm(CourseType::class);
         $form->handleRequest($request);
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-            return $this->redirectToRoute('course_index');
+            $formData = $form->getData();
+            try {
+                $editResponse = $billingClient->updateCourse($formData, $slug, $this->getUser()->getApiToken());
+            } catch (HttpException $e) {
+                return $this->render('course/edit.html.twig', array(
+                    'course' => $combinedCourse,
+                    'form' => $form->createView(),
+                    'error' => $e->getMessage()
+                ));
+            }
+            if (array_key_exists('code', $editResponse)) {
+                return $this->render('course/edit.html.twig', array(
+                'course' => $combinedCourse,
+                'form' => $form->createView(),
+                'error' => $editResponse['message']
+                ));
+            } else {
+                $course->setName($formData['name']);
+                $course->setDescription($formData['description']);
+
+                $entityManager = $this->getDoctrine()->getManager()->flush();
+                return $this->redirectToRoute('course_index');
+            }
         }
         return $this->render('course/edit.html.twig', [
-            'course' => $course,
+            'course' => $combinedCourse,
             'form' => $form->createView(),
+            'error' => null
         ]);
     }
 
@@ -90,12 +134,28 @@ class CourseController extends AbstractController
      * @Route("/{slug}", name="course_delete", methods={"DELETE"})
      * @IsGranted("ROLE_SUPER_ADMIN")
      */
-    public function delete(Request $request, Course $course): Response
+    public function delete($slug, Request $request, Course $course, BillingClient $billingClient): Response
     {
         if ($this->isCsrfTokenValid('delete'.$course->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($course);
-            $entityManager->flush();
+            $foundCourse = $this->getDoctrine()->getRepository(Course::class)->findOneCombined($slug, $billingClient, $this->getUser());
+            try {
+                $deleteResponse = $billingClient->deleteCourse($slug, $this->getUser()->getApiToken());
+            } catch (HttpException $e) {
+                return $this->render('course/show.html.twig', array(
+                    'error' => "Сервис временно недоступен. Попробуйте отредактировать урок позднее",
+                    'course' => $foundCourse
+                ));
+            }
+            if (array_key_exists('code', $deleteResponse)) {
+                return $this->render('course/show.html.twig', array(
+                'course' => $course,
+                'error' => $deleteResponse['message']
+                ));
+            } else {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($course);
+                $entityManager->flush();
+            }
         }
         return $this->redirectToRoute('course_index');
     }
@@ -106,7 +166,14 @@ class CourseController extends AbstractController
      */
     public function buyCourse($slug, BillingClient $billingClient, CourseRepository $courseRepository): Response
     {
-        $result = $billingClient->buyCourse($slug, $this->getUser()->getApiToken());
+        try {
+            $result = $billingClient->buyCourse($slug, $this->getUser()->getApiToken());
+        } catch (HttpException $e) {
+            $this->addFlash('error', 'Сервис временно недоступен. Попробуйте удалить урок позднее"');
+            return $this->render('course/show.html.twig', [
+                'course' => $courseRepository->findOneCombined($slug, $billingClient, $this->getUser())
+            ]);
+        }
         if (array_key_exists('success', $result)) {
             $this->addFlash('success', 'Курс успешно оплачен');
             return $this->render('course/show.html.twig', [
